==> 05_analyses/04_msit_dmn_task.Rmd <==
# Association between MSIT task design and DN signal

Note that actual code is loaded from a different file.

```{r external-code, cache=FALSE}
read_chunk('04_msit_dmn_task.R')
```

## Setup

```{r general-setup, message=FALSE, results='hide'}
```

```{r network-setup}
```

```{r subject-info}
```

Note that negative signal here are the congruent trials and positive signal are the incongruent trials. What do you think would happen if the positives and negatives got together?

```{r task-waver, fig.width=8, fig.height=4}
```

```{r rts}
```

```{r msit-ts}
```

## Group Average

Get the xmin and xmax for showing blocks.

```{r msit-design}
```

### BOLD Time Series

Compute the group average MSIT BOLD signal.

```{r msit-dmn-average}
```

Let's see you work it ggplot. Shake it. Shake it good.

```{r msit-dmn-average-plot, fig.width=12, fig.height=4, dev=c("png","cairo_ps")}
```

### RT Time Series

Compute the group average RT 'signal'.

```{r msit-rt-average}
```

Let's see you work it ggplot. Shake it. Shake it good.

```{r msit-rt-average-plot, fig.width=12, fig.height=4, dev=c("png","cairo_ps")}
```

### BOLD & RT Time Series

Let's see you work it ggplot. Shake it. Shake it good.

```{r msit-dmn-rt-average-plot, fig.width=12, fig.height=4, dev=c("png","cairo_ps")}
```

### Correlations Between DN and RT Signal

```{r msit-dmn-rt-correlation}
```

## Correlations Between Task Design and DN Signal

```{r msit-dmn-correlation}
```

```{r combine}
```

```{r functions-phenotype}
```

### Total Scale Scores (with BDI, without PANAS)

#### Multiple Regression

```{r multiple-totals-correlation-bdi, fig.width=12, fig.height=4}
```

#### Single Regressions

```{r single-totals-correlation, fig.width=12, fig.height=4}
```

### Total Scale Scores (with PANAS, without BDI)

#### Multiple Regression

```{r multiple-totals-correlation-panas, fig.width=12, fig.height=4}
```

#### Single Regressions

Of course there is no point in re-running this here.

### RRS SubScales

#### Multiple Regression

```{r multiple-rrs-correlation, fig.width=10, fig.height=4}
```

#### Single Regressions

```{r single-rrs-correlation, fig.width=10, fig.height=4}
```

### SIPI SubScales

#### Multiple Regression

```{r multiple-sipi-correlation, fig.width=8, fig.height=4}
```

#### Single Regressions

```{r single-sipi-correlation, fig.width=8, fig.height=4}
```

### ERQ SubScales

#### Multiple Regression

```{r multiple-erq-correlation, fig.width=8, fig.height=4}
```

#### Single Regressions

```{r single-erq-correlation, fig.width=8, fig.height=4}
```

### PANAS SubScales

#### Multiple Regression

```{r multiple-panas-correlation, fig.width=8, fig.height=4}
```

#### Single Regression

```{r single-panas-correlation, fig.width=8, fig.height=4}
```
